refactor(web): extract shared collaborator modal helpers
Deduplicate normalizeUsername, rejectedItems, CollaboratorIdentity, and the
common GitHubAPIError failure mapping across the three collaborator dialogs
into components/modals/collaboratorHelpers.tsx, and use the RepoPermission type
in place of the repeated permission literal union. Behavior-preserving.
